+++ src/useSelector.js
@@ -1,13 +1,6 @@
-import {
-  useContext,
-  useEffect,
-  useRef,
-  useReducer,
-} from 'react';
+import { useContext, useEffect, useReducer } from 'react';
 
 import { defaultContext } from './Provider';
-
-import { useIsomorphicLayoutEffect } from './utils';
 
 const isFunction = (f) => typeof f === 'function';
 const defaultEqualityFn = (a, b) => a === b;
@@ -17,33 +10,22 @@ export const useSelector = (selector, eqlFn, opts) => {
     equalityFn = isFunction(eqlFn) ? eqlFn : defaultEqualityFn,
     customContext = defaultContext,
   } = opts || (!isFunction(eqlFn) && eqlFn) || {};
-  const [, forceUpdate] = useReducer((c) => c + 1, 0);
   const { state, subscribe } = useContext(customContext);
-  const selected = selector(state);
-  const ref = useRef(null);
-  useIsomorphicLayoutEffect(() => {
-    ref.current = {
-      equalityFn,
-      selector,
-      state,
-      selected,
-    };
-  });
+  const [selected, updateSelected] = useReducer((prevSelected, nextState) => {
+    const nextSelected = selector(nextState);
+    if (equalityFn(prevSelected, nextSelected)) return prevSelected;
+    return nextSelected;
+  }, state, selector);
+  let selectedToReturn = selected;
+  const currSelected = selector(state);
+  if (!equalityFn(selected, currSelected)) {
+    // schedule another update, because state from context has been changed
+    updateSelected(state);
+    selectedToReturn = currSelected;
+  }
   useEffect(() => {
-    const callback = (nextState) => {
-      try {
-        if (ref.current.state === nextState
-          || ref.current.equalityFn(ref.current.selected, ref.current.selector(nextState))) {
-          // not changed
-          return;
-        }
-      } catch (e) {
-        // ignored (stale props or some other reason)
-      }
-      forceUpdate();
-    };
-    const unsubscribe = subscribe(callback);
+    const unsubscribe = subscribe(updateSelected);
     return unsubscribe;
   }, [subscribe]);
-  return selected;
+  return selectedToReturn;
 };